requirements.txt
# Django itself
django<3.3,>=3.2

# Helpers
django-environ==0.8.1
whitenoise==6.0.0
django-crispy-forms==1.14.0
requests>=2,<3


# Used in development only
django-debug-toolbar==3.2.4
pre-commit==2.17.0
black==22.1.0
reorder-python-imports==3.0.1
tox==3.24.5


# Label generator
git+https://github.com/davidfischer/mtg-printable-set-label-generator.git
